+++ test_PDA.py
@@ -3,14 +3,13 @@ from base import RuleBook, MultiRuleBook
 
 
 def test_DPDA():
-    import DPDA
     rulebook = RuleBook([
         PDARule(1, '(', 2, '$', ['b', '$']),
         PDARule(2, '(', 2, 'b', ['b', 'b']),
         PDARule(2, ')', 2, 'b', []),
         PDARule(2, None, 1, '$', ['$'])
     ])
-    dpda_design = DPDA.DPDADesign(1, [1], rulebook)
+    dpda_design = PDADesign(1, [1], rulebook)
     assert dpda_design.is_accepted('(((((((((())))))))))') == True
     assert dpda_design.is_accepted('()(())((()))(()(())))') == False
     assert dpda_design.is_accepted('(()(()(()()(()()))()') == False
@@ -18,7 +17,6 @@ def test_DPDA():
 
 
 def test_DPDA_2():
-    import DPDA
     rulebook = RuleBook([
         PDARule(1, 'a', 2, '$', ['a', '$']),
         PDARule(1, 'b', 2, '$', ['b', '$']),
@@ -28,14 +26,13 @@ def test_DPDA_2():
         PDARule(2, 'b', 2, 'a', []),
         PDARule(2, None, 1, '$', ['$'])
     ])
-    dpda_design = DPDA.DPDADesign(1, [1], rulebook)
+    dpda_design = PDADesign(1, [1], rulebook)
     assert dpda_design.is_accepted('ababab') == True
     assert dpda_design.is_accepted('bbbaaaab') == True
     assert dpda_design.is_accepted('baa') == False
 
 
 def test_NPDA():
-    import NPDA
     rulebook = MultiRuleBook([
         PDARule(1, 'a', 1, '$', ['a', '$']),
         PDARule(1, 'a', 1, 'a', ['a', 'a']),
@@ -59,7 +56,6 @@ def test_NPDA():
 
 
 def test_NPDA_2():
-    import NPDA
     rulebook = MultiRuleBook([
         PDARule(1, 'a', 1, '$', ['a', '$']),
         PDARule(1, 'a', 1, 'a', ['a', 'a']),
@@ -74,7 +70,7 @@ def test_NPDA_2():
         PDARule(2, 'b', 2, 'b', []),
         PDARule(2, None, 3, '$', ['$']),
     ])
-    npda_design = NPDA.NPDADesign(1, {3}, rulebook)
+    npda_design = PDADesign(1, {3}, rulebook)
     assert npda_design.is_accepted('') == True
     assert npda_design.is_accepted('abba') == True
     assert npda_design.is_accepted('babbbbab') == True
